Dependency uptake in prodcon
To get a coherent set of packages from the prodcon build, we need
the package reference from Microsoft.Build.Runtime to be overrideable
in a prodcon environment.

This commit separates the SDK version we're using to run the build from
the netcoreapp package version we're referencing, so that the latter
can be overridden by repo-toolset tooling in prodcon.