Never mark files from nuget as in-box fx ref assemblies
Before conflict resolution between nuget and redist list, we would end up in
situations where MSBuild marked an assembly from nuget matching redist list with
<= version with FrameworkFile=true metadata. This caused us to designate the
assembly as type "referenceassembly" in deps.json for
PreserveCompilationContext. However, that is not correct as that designation
means that the assembly is literally from the targeting pack in Program Files
and not from nuget.

This no longer happens because the nuget assembly with smaller assembly version
will be discarded during conflict resolution, favoring the in-box assembly that
is correctly designated as "referenceassembly". Furthermore, an OOB with greater
version than in-box will not get FrameworkFile=true set by MSBuild. As such,
there's no longer a known real scenario (and hence no test here) where we will
observe FrameworkFile=true on an assembly from nuget. This is therefore an
defensive change in case it ever happens in the wild for unforeseeable reasons.