Avoid handshake timeout on Windows
The old version of this code didn't have a timeout when reading the
handshake packet, and appeared to be reliable. The timeout was required
to work around a cross-platform design limitation: the pipe always
connects on *nix systems, even when nothing is listening to it.

This caused failures on Windows (#3214), because the timeout was
exceeded there. For maximum robustness, avoid the timeout entirely on
Windows.