[test] Fix test clean up failures for CommentsInPreprocessing
    Microsoft.Build.UnitTests.UtilitiesTestStandard.CommentsInPreprocessing

This test depends on `-pp`/preprocessing and directly invokes
`MSBuildApp.Execute`, which sets the environment variable
`MSBUILDLOADALLFILESASWRITEABLE` and never unsets it.

`TestEnvironment`+`MSBuildTestEnvironmentFixture` collects the set of
environment variables before the test begins but finds the extra
`MSBUILDLOADALLFILESASWRITEABLE` at the end causing a failure in the
test cleanup. This cascades into a test-class cleanup failure and on to
an assembly level cleanup failure.

This shows up as xunit `Error` on netcore and as a `Failed` on mono, not
sure why exactly though.