Extend handshake-response timeouts on *nix
On non-Windows OSes, use preexisting timeouts when waiting for responses
after connecting to the IPC pipe. Instead of 0ms, passed when polling
through preexisting nodes trying to reuse one, enforce a minimum timeout
of 50ms. This is unfortunately a pure guess--the 30ms timeout that was
used before was cribbed from the timeout in `ShutdownAllNodes`, but
seemed too short given failures on heavily-loaded machines in aspnet
universe. However, the scenario there is connecting to just-launched
nodes, so will be subject to a 30s (`TimeoutForNewNodeCreation`) time,
so won't be subject to the minimum.