[test] Remove an unnecessary test
RAR_Tests.SubsetListFinderVerifyEmptyInSubsetsToSearchForAndCaching

This is essentially checking if the the results of two different queries
are two *different* empty array objects. This is an implementation
detail and should not be relied upon. It is a legal optimization to
return the same empty array object, which is what List.ToArray does on
netcore and mono (on macOS). Full framework returns the same object, so
this test passed on .net/windows.

(cherry picked from commit 9f3bffbc9976ac93bd6d45e4fe417757d2e80f1c)