Avoid explicitly specifying .NET Core version
We took an explicit dependency on Microsoft.NETCore.App in e735b93 to
ensure that we would reference final/stabilized versions of the package
without having to commit a change ourselves. But the NuGet reference to
that package isn't required; listing it was dotnet/sdk#2204.

This commit returns to referencing Microsoft.NETCore.App implicitly,
and includes a workaround to keep its version from affecting our
output packages.