[test] Remove an invalid check from the test ..
..  RAR_Tests.SubsetListFinderVerifyEmptyInSubsetsToSearchForAndCaching

This is essentially checking if the the results of two different queries
are two *different* empty array objects. This is an implementation
detail and should not be relied upon. It is a legal optimization to
return the same empty array object, which is what List.ToArray does on
netcore and mono (on macOS). Full framework returns the same object, so
this test passed on .net/windows.

So, remove that part of the test and reduce it to just ensuring that the method
does not crash if it is passed empty string elements in the array.
And rename to ..

    RAR_Tests.SubsetListFinderVerifyEmptyInSubsetsToSearchFor