Remove WaitForDebugger from our "public" API
This property is public, but only exists in debug builds, which are
never "official" builds (not distributed in NuGet packages or in VS).

This also caused a builds run with `-configuration Release` to edit the
API files, removing WaitForDebugger.

Remove it from the API permananently and tell GenAPI to stop adding it.