Add support for fallback paths to be specified via property references
Currently, the list of fallback paths in app.config can have a property
reference, like:

```
 <searchPaths os="osx">
	<property name="MSBuildExtensionsPath" value="$(FallbackPath);/x/y/z" />
```

.. but it is assumed that such a property will expand to a single path.

This commit adds support for such properties to expand to multiple ';'
separated list of paths. The rules for expansion do not change, they are still
expanded at the time of project import.

With this, given `$(FallbackPath)="/a/b/c;/q/w/e"` and the list of fallback paths
would become `"/a/b/c;/q/w/e;/x/yz"`.