[mono][osx] Add default fallback search paths for MSBuildExtensionsPath
On Mono, $(MSBuildExtensionsPath) is $mono_prefix/lib/mono/xbuild.
On OS X, this is:

	`/Library/Frameworks/Mono.framework/Versions/Current/lib/mono/xbuild`

Consequently, when resolving an Import for
$(MSBuildExtensionsPath)\Foo\Abc.targets, the following path will be checked:

	`/Library/Frameworks/Mono.framework/Versions/Current/lib/mono/xbuild/Foo/Abc.targets`

This works, but has a problem when upgrading the Mono.framework package:
a *new* /Library/Frameworks/Mono.framework/Versions/@VERSION@ directory
is created, so anything installed into the previous directory will be
"lost". The file is still there, but xbuild will no longer find it.

To address this upgrade scenario, Mono's msbuild/xbuild also check a fallback path:

	`/Library/Frameworks/Mono.framework/External/xbuild`

This is used by Xamarin products (Xamarin.iOS/Mac, Xamarin.Android)