Complete submission logging on exception
When the scheduler completes a request, it looks ahead at the queue of unscheduled requests and
preemptively completes identical ones (same global properties, path, toolsversion, AND requested targets).
When the BuildResult has a result for the unscheduled identical request (most of the times), then the scheduler
simulates the ProjectStarted / ProjectEnded events for the unscheduled request and completes it.

However, when the initial build fails in such a way that there is no result for the requested targets,
the initial BuildResult cannot satisfy the unscheduled submission. In this scenario the Scheduler still completes
the request but does not simulate the ProjectStarted / ProjectEnded events. It also leaves logging completion to the
BuildManager.

This hangs the build, as nobody is marking the request's submission as completed. The lowest risk fix, done here,  is to have the BuildManager complete logging whenever there is an exception.

A medium risk change would have been to have the Scheduler simulate started / ended events for these lookahead requests with exceptions. A high risk change would have been to change the TargetBuilder to include results for targets with missing target dependencies.