Add ZipDirectory and Unzip tasks (#3291)
`ZipDirectory` task zips up a whole directory.  It will fail if the destination file already exists.  Zipping individual files is a lot harder because you have to calculated a base path for all files and make zip entries relative to that.  The code in the CLR that does this is very efficient by reusing string buffers and I didn't want to have to replicate the code.  So for now there's only `ZipDirectory` instead of a more generic `Zip`.

`Unzip` unzips files from an archive to a directory.
* Defaults to skip unzipping files that are already up-to-date
* Logs every file that was unzipped
* Supports cancellation
* Supports overwriting read-only files
* Supports unzipping multiple files to the same directory

Fixes #1781