Merged PR 126122: Fix alpine file ownership issues with newer docker
Fix alpine file ownership issues with newer docker
Upgrades to a newer docker version (18.03.1-ce) caused files created inside to be owned by root on alpine.  It appears that the logic to set up the user in the container so that this doesn't happen was missing from alpine.  While it's not clear why it worked before at all, the logic has been duplicated (tweaked for the alpine base image).