Prevent copying removed generated content
When using the `PrepareForPublish` hook, extra content (CSS, JS, etc.) is often generated to be published with the project. However, sometimes the steps taken during this hook will also cause content that was previously generated to be removed. For instance, a JS library may have been used but is not used anymore, so a `gulp` task may remove it from `wwwroot/lib`. When this happens, the SDK should respond appropriately or else it will try to copy non-existing files to the publish output directory, resulting in build-halting errors. This commit updates the `_IncludePrePublishGeneratedContent` target to do just that.