Remove usage of MSBUILDRELOADTRAITSONEACHACCESS environment variable (#3385)
Traits: This caches values of various environment variables which act as
switches to control features like logging imports. For use with
tests, this caching can be disabled but that is enabled only for DEBUG
builds. Which means that any tests trying to toggle and test those
environment variable dependent features will work only for debug builds.

This is also controlled by the environment variable,
`MSBUILDRELOADTRAITSONEACHACCESS`, which is no longer required.

@cdmihai: I remember that in the past, we relied on the env var as a
way for tests to communicate to the Traits to refresh themselves. We
needed the var because Traits is compiled into every assembly, so every
assembly has it's own copy of the object, and we had no easy way of
addressing all of them. But now we use reflection as a single source of
truth, which makes me believe we don't need the env var at all.

(https://github.com/Microsoft/msbuild/pull/3385#issuecomment-397376338)