Fix how warnings are mutated
The list of warnings to mutate are stored according to the project instance ID.  When a project is finished, the warnings are removed.  However, a project starts and finishes several times during a build.  This leads to warnings failing to elevated or suppressed.

This fix moves the warning mutation logic out of EventSourceSink into the LoggingService itself instead.  This simplifies the implementation a little.

It also changes the key to the project instance ID and project context ID.  When projects start and finish, the key will work properly.

Fixes #3474