Ensure DefaultProjectTypeGuid is set when cross-targeting.
For C# and VB projects that are cross-targeted, the cross-targeting targets are
evaluated from the root project and not the "current version" targets for the
languages (this is currently where `DefaultProjectTypeGuid` is set for each
language).  The "current version" targets are evaluated only from the "inner"
builds perform for each framework target; thus the `DefaultProjectTypeGuid` is
never set from the evaluation of a cross-targeted project file.

This means that when `dotnet` evaluates the project to determine the default
project type GUID to use for a command like `dotnet sln add`, it fails to do
and displays an error message to the user.

See related CLI issue dotnet/cli#9477.

The fix is to duplicate setting `DefaultProjectTypeGuid` in the cross-targeting
targets.  This allows the property to be visible when evaluating the root
project file.