Report TFM via GetTargetFrameworks if TF undefined
Prior to this change, ProjectReferences to non-SDK projects did not
report a `TargetFrameworks` list to the referencing project because
prior to the .NET Core SDK, the TFM for a project was defined in
`$(TargetFrameworkMoniker)` as the long form of the moniker.

Fixes #3546 by falling back to the legacy TFM in the included-in-all-
projects copy of `GetTargetFrameworks`.