Implement a workaround for DefaultProjectTypeGuid and cross-targeting.
Currently MSBuild only sets the `DefaultProjectTypeGuid` property from the
language properties for a project that is not cross-targeting (it gets set on
the inner projects that are invoked for each target framework instead).  As a
result, `dotnet sln add` is currently not working for projects that
cross-target.

This commit implements a workaround until such a time that MSBuild can ensure
`DefaultProjectTypeGuid` is set via the language-specific cross-targeting
targets.

See https://github.com/Microsoft/msbuild/issues/3558 for the issue tracking the
fix on the MSBuild side.

Fixes dotnet/cli#9477.