[mono] Update NuGet package version to match what we have in the sdks
.. in mono/buid/SdkVersions.txt . We need to keep these two in sync and
usually they shouldn't be too far apart. In future, the way we build
mono/msbuild will change so that this won't be an issue at all.

We could have read the version from the txt file and enforced it via the
property, `$(NuGetPackageVersion)`, but I'm choosing to edit the props
file so that we get a conflict whenever we merge from upstream and don't
forget to pay attention to the version update, in case that requires any
additional work/testing.