Fix setting SelfContained to false.
The 2.1.400 SDK changed the way `SelfContained` was being set: it would only set
it to `true` if not set, otherwise it would leave it as an empty string.

While this was fine for the .NET Core SDK because all "false" checks against
`SelfContained` were using `'$(SelfContained)' != 'true'`, thus that expression
would evaluate to true when `SelfContained` was empty.

However, the 2.0 Web SDK used an expression of `'$(SelfContained)' == 'false'`,
which no longer evaluated to true because `SelfContained` was now an empty
string.  This resulted in publishing a project as self-contained when it should
have been framework-dependent.

The fix is to set `SelfContained` to false like the .NET Core SDK was doing
before.

Likewise, this commit also sets the related `UseAppHost` property to false when
it would be empty otherwise in case external SDKs make similar comparisons.

Fixes dotnet/cli#9852.