Add resolver cache
It is used to speed up local tools command resolution. Currently is an isolated component.

Correctness:
In this version, the correctness is the first. And there are a lot of usage of linq and creating new objects. However, from the profile of the prototype, reading from the cache is trivial compared with other places. Also, there is no end to end code, so I hope to optimize this part after I have it. At the same time, there are many strategies for different bottle necks if it shows up. Like, I can order the cache during save or add an extra field to store the hash of fields. If we spend more time reading manifest file, we should spend more development effort to cache manifest file info instead.

When it will be used:
After user issue tool restore, we will run NuGet restore, and we can get IReadOnlyList<CommandSettings> as the result. Cache Save will be run to store the mapping between Package Id–Version-TFM-RID and tools command available. Load will be used during command resolution or picking out already restored commands.

Concurrency:
No lock on write, if file is corrupted treat it as empty. And a new restore will be triggered and the cache will be rewrite if it is during tool restore.

The reason being, there are only 2 ways a file could be corrupted. 1. two CLI running at the same time. Since there should not be duplication in package id in one manifest file (the code to validate is coming). And in this case, it is user error. 2. Direct edit the cache, it is still a user error.

In the current design, if above happens, we can still get back to correct state. If we use a more complex read and write lock with a file. We could get in a nastier bug of locking file forever for a potentially non-exist file corruption case.

What on disk:
FileName: my.toolBundle
[
    {
        "Version": "1.0.2",
        "TargetFramework": "netcoreapp2.1",
        "RuntimeIdentifier": "any",
        "SerializableCommandSettingsArray": [
            {
                "Name": "tool1",
                "Runner": "dotnet",
                "RelativeToNuGetGlobalPackagesFolderPathToDll": "tool1.dll"
            },
            {
                "Name": "tool2",
                "Runner": "dotnet",
                "RelativeToNuGetGlobalPackagesFolderPathToDll": "tool2.dll"
            }
        ]
    }
]