Launch with apphost for `dotnet run`.
This commit changes `dotnet run` to respect the runtime specified via the
`--runtime` option when querying the project for the run command and arguments.

By doing so, the run command will change to the apphost instead of using
`dotnet`.  Thus, the apphost gets executed instead of `dotnet <dll>`.

Additionally, this commit adds the `-r` short form of the `--runtime` command
so that it is consistent with the other commands.

Fixes #9866.