Separate _CreateAppHost from _ComputeNETCoreBuildOutputFiles
This cuts the dependency from _ComputeNETCoreBuildOutputFiles on
CoreCompile. Instead, _CreateAppHost will run after compilation via
CompileDependson.

Locating the restored apphost has been factored into another target
_GetAppHostPaths, which is a dependency of both
_ComputeNETCoreBuildOutputFiles and _CreateAppHost. The destination
apphost path is also computed here, and passed to the CreateAppHost
task as an input.

This way, during a "publish --no-build", the apphost destintation
path (of the previously-generated apphost) is computed as a dependency
of _ComputeNETCoreBuildOutputFiles, and it is copied to the publish
directory without causing CoreCompile to run.