Fix apphost incremental build
The apphost is (optionally) customized by copying resources from the
managed dll, and setting the GUI bit based on the Target
property (which also modifies the managed dll). Therefore, we use the
managed dll as an input for incremental build (the apphost needs to be
updated iff the managed dll was updated). This fixes
https://github.com/dotnet/sdk/issues/2554.

This also removes the unnecessary overwriteExisting flag from
AppHost.Create, fixing https://github.com/dotnet/sdk/issues/2473.