Fix conflict resolution test
This test was previously incorrect.  It was using the wrong property name to try to disable version information
from being written to the deps file.  This meant that version information was being written to the deps file.
Furthermore, the test expectation was backwards from the test name: the test was erroneously expecting
the DLL to be loaded from the app's publish directory.

The broken version of the test started failing once we got a stage 0 that included this fix:
https://github.com/dotnet/core-setup/pull/4521.  That PR changed the behavior when the versions of
the file in the published app and the runtime matched.  In that case, the runtime will now prefer the
runtime copy instead of the app's copy.