Ensure default apphost RID is set prior to `_IsExecutable` reassignment.
Because of the `_IsExecutable` reassignment that happens when the Test SDK
changes `OutputType`, a test project that builds with `HasRuntimeOutput` set to
true will cause a build failure due to the default apphost not being restored,
but requested to be resolved as a project asset.

The fix is to ensure that the `DefaultAppHostRuntimeIdentifier` property is set
prior to reassigning the `_IsExecutable` property.

Fixes #2620.