Fix copying of build assets when publish assets differ.
When publish copy-local assets differ from build copy-local assets (for
example, runtime store packages were excluded), a typo when adding the
reference copy local items caused items that were copied local for build to be
included in publish output.

This was caught by a CLI test that uses a manifest to exclude a package and
then runs the output of the publish operation, expecting it to fail because the
dependency was not copied locally (and no runtime store is used).

This change will fix the failing CLI test.