Fix the serialization of ProjectItemInstance
Fixes [Bug 68841](https://devdiv.visualstudio.com/DevDiv/_workitems/edit/688412)/[Feedback 678764](https://devdiv.visualstudio.com/DevDiv/_workitems/edit/678764).

Fix the serialization of `ProjectItemInstance` such that `_itemDefinitions` (a `List<ProjectItemDefinitionInstance`) and `_directMetadata` (a `CopyOnWritePropertyDictionary<ProjectMetadataInstance>`) are serialized separately.

Currently, when serializing a `ProjectItemInstance` we collapse the metadata in both `_directMetadata` and `_itemDefinitions` into a single dictionary via the `MetadataCollection` property (when deserializing these fields will be null and/or empty collections at this point, of course). Then, when serializing _or_ deserializing we set `_directMetadata` to this new dictionary.

That means that if you start with a `ProjectItemInstance` with non-empty collections for `_itemDefinitions` and `_directMetadata` then after the serialization it will have the same `_itemDefinitions` but the `_directMetadata` will contain all the items that were previously in both `_itemDefinitions` _and_ `_directMetadata`. After deserialization you'll end up with a `ProjectItemInstance` with a null/empty `_itemDefinitions`, and `_directMetadata` will have the combined items.

As best I can tell this is simply wrong. Problems can crop up because direct metadata is already evaluated, while the metadata from item definitions is not. So after one of these serialization/deserialization cycles we will have items in `_directMetadata` that we assume are evaluated, but are not.

This manifests in various ways, from unexpanded/unevaluated text showing in MSBuild messages, to generating files with names like "%(BaseFileName)PerfCounters.g.cs".

The fix here is to keep `_itemDefinitions` and `_directMetadata` separate during the serialization/deserialization process.