Use `PreserveNewest` for explicitly published files.
The publish targets contain files that are explicitly published, such as the
build output, the deps file, the runtime config, etc.  Because the default for
resolved files to publish was to always copy, a second publish operation when
everything was up-to-date had needless copies of these files.

This commit changes those items to use `PreserveNewest` so that they are only
copied if the source is newer than the target.

Fixes #2573.