Skip resource updating on Nano Server.
This commit fixes the resource updater usage to be skipped when building on
Nano Server.

For Windows Server Nano 2016, this fixes an `EntryPointNotFoundException` because
`BeginUpdateResource` cannot be found in kernel32.

For 1709+ of Windows Server Nano, this fixes an `HResultException` with code
`80070000`, because the export resolves but seemingly is not implemented.

Fixes #2652.