Improve Ctrl-C experience for `dotnet run`.
This commit improves the handling of the SIGINT signal from `dotnet run`.

The current behavior of `dotnet run` is to use the default handler for the
SIGINT signal.  This results in the termination of the `dotnet run` process
with an exit status of 130 on POSIX systems and 0xC000013A on Windows.  This
prevents a child process to gracefully handle SIGINT by returning a zero exit
code or from ignoring the signal entirely and letting the process appear to
continue to run.

Additionally, the child process of `dotnet run` may still be writing to stdout
during the handling of SIGINT.  Because `dotnet run` may terminate before the
child process does, a waiting parent of `dotnet run`, such as a shell, may
proceed to print out a new prompt to the user while the child continues to
write to stdout.  This results in inconsistent and interleaved output in a
terminal and can be a source of confusion to users.

This commit changes `dotnet` to always ignore SIGINT when spawning a child
process.  This allows the child process to handle the signal and decide what to
do with it.  As a result, the UX from when the user program is run directly and
run via `dotnet run` is the same regarding SIGINT.

Fixes #812.