Add a new intrinsic item function
Add a new intrinsic item function, `GetPathsOfAllFilesAbove`, to get all the files of a given name that are in the same or ancestor directories of the given items. This is being done to support the ".editorconfig in compiler" scenario where we need to map from all the .cs/.vb files in the project to the applicable set of .editorconfig files. For various reasons we don't want to find the .editorconfig files using a target/task, so instead we are adding this intrinsic function to allow us to find them as part of project evaluation.

There are two phases to the implementation. First we build the set of directories that we will search. For each input item, we get the item's directory and see if it is already in the set. If it is we don't need to do anything further with that item. Otherwise we walk up from that directory to the root, adding each ancestor directory to the set (or bailing out early if it has already been added).

Second, we go through the set of directories looking for the specified file. If it is found we return it.

This way we only check for each potential file once, rather than once per input file under that directory.