Remove unit test assert
Remove an unnecessary unit test assert. The `Assert.Collection` call in `ExpandItemVectorFunctions_GetPathsOfAllFilesAbove` already checks the complete contents of the collection; there's no need for a separate assert for the item count.

This unit test is failing on Mac and Linux because we're only finding one path where we expect two; by removing this assert I hope to see which path is missing.