Fix NativeMethodsShared.GetLastWriteFileTimeUtc when the path is a
.. directory.

`GetLastWriteFileTimeUtc` depends on the incorrect assumption[1] that
`File.GetLastWriteTimeUtc` will return `DateTime.FromFileTimeUtc(0)` if
the path is a directory.

The same thing applies for the `Directory` equivalent of the same
method.

This breaks File existence checks, for example in
`ResolveAssemblyReference` when we have a reference like:

`<Reference Include="FooBar" />`

.. where `FooBar` is a directory in the current(project) directory.
ResolveAssemblyReference ends up triyng to get assembly information from
the file, because it thinks that `FooBar` is a file and thus an assembly
path.