Limit default apphost feature to netcoreapp3.0 targeted apps.
This commit limits the default apphost feature to applications that target
`netcoreapp3.0` or later.

This prevents altering the expected outputs of previous projects after an
upgrade to a 3.0 SDK.

More importantly, this allows for the removal of the hack that was put in place
to force a restore of the SDK RID to obtain the apphost; the apphost is now
resolved through framework references.

Fixes dotnet/cli#10566.