cibuild_bootstrapped_msbuild.sh: Build stage1 and boostrapped in fixed
.. directories.

We were building in `artifacts/` for stage1 and then moved it to
`stage1/`, and the bootstrapped build was built in `artifacts/`. IIUC,
this is because the `eng/common/tools.sh` didn't allow overriding the
`artifacts_dir` which was changed in a previous commit.

Now, stage1 will build in `artifacts` directory and the bootstrapped
build in `artifacts/2` as before.

Also, pass configuration to the arcade scripts.