Don't set MSBuildAllProjects
Microsoft/msbuild#1299 discusses a change whereby MSBuild prepends the newest input file path to this property automatically. Therefore, so long as consumers of this property only use this list for up-to-date checks, it's possible to reduce the size of this property, and reduce corresponding allocations.

This reduction applies equally to consumers, such as in the case discussed in dotnet/project-system#3744

For backwards compatibility the property is still set for MSBuild versions prior to 16.0.