Fix ArgumentNullException in ResolveFrameworkReferences.
When a `netcoreapp3.0` executable project references a `netcoreapp3.0`
library project, and `SelfContained` is globally true, then an
`ArgumentNullException` occurs in the `ResolveFrameworkReferences` task.

This can happen when publishing with the following command:

```
dotnet publish -f netcoreapp3.0 -r win-x86 --self-contained
```

When `SelfContained` is `true` for the class library project reference,
the task does not take into account that `RuntimeIdentifier` might be
null since it gets filtered out from the reference.

Additionally, some cleanup was needed relating to
`GetBestRuntimeIdentifier`.

Fixes #2851.