Initial work to enable OptProf to run.
Removing vswhere.exe and adding .tools to .gitignore.

Changing eng/build.ps1 to be a super set of eng/common/build.ps1. That's the guidance I got from Tomas.

Deleting the publish test results steps for official builds. We are not running tests there.

Update the official build yaml to publish optprof data and upload to vs drop with the right container name.