Fix _RestoreCrossgen target dependency.
The `_RestoreCrossgen` target depends on a package reference to
`Microsoft.NETCore.App` to be present.  However, for `netcoreapp3`
targeted applications, this requires the `ResolveFrameworkReferences`
target to execute.

This commit adds a target dependency from `_RestoreCrossgen` to
`ResolveFrameworkReferences` so that the package references are
populated when coming from a framework reference.

Fixes the CLI tests around `dotnet store` that invoke the crossgen
targets.