Add test for ComposeStore target for netcoreapp3.0.
This commit adds a test to cover using the `ComposeStore` target from a
project targeting `netcoreapp3.0`.

The test is intended to replicate similar tests we have in `dotnet/cli`
so that we can catch regression of the `dotnet store` command in the SDK
repo.

The test ensures `ComposeStore` will work when targeting `netcoreapp3.0`
and that the underlying issue (a reference to `Microsoft.NETCore.App`
being added without a version) is checked for.