Ensure Microsoft.NETCore.App is an implicit dependency for publish.
This commit ensures that we exclude files from the `Microsoft.NETCore.App`
platform library on publish when the platform library is
`Microsoft.AspNetCore.App`, which doesn't have an explicit dependency on
`Microsoft.NETCore.App`.

When publishing a 2.x ASP.NET application, the `MicrosoftNETPlatformLibrary`
property gets changed by ASP.NET to `Microsoft.AspNetCore.App`.  This causes
the task responsible for determining what files to copy locally to treat files
from `Microsoft.NETCore.App` and its dependencies as not being part of the
platform.

However, when publishing the application as framework-dependent, almost all the
platform files get excluded thanks to conflict resolution which has the files
as part of the platform manifest and preferred packages.  Unfortunately, the
apphost, hostpolicy, and hostfxr files are not part of these lists and as a
result will be copied locally.

This breaks framework-dependent apphost activation because hostfxr and
hostpolicy are copied locally.  Additionally, these files end up in the
deps.json file which also prevents roll-forward activation if the two files are
manually deleted.

The fix is to treat `Microsoft.NETCore.App` as an implicit dependency of
`Microsoft.AspNetCore.App` where we calculate the list of excluded packages for
a framework-dependent publish.

Fixes dotnet/cli#10602.