Allow WriteLinesToFile to create empty files (#4125) (#4130)
A change intended as a performance optimization in #4067 caused a
backward-compatibility regression: `WriteLinesToFile` became unable to
create empty files. `Touch` would be more appropriate for that case, but
`WriteLinesToFile` is used in that way in the Framework-delivered
`GenerateCompiledExpressionsTempFile` target, which we can't change
easily.

Removes the condition that caused us to avoid calling `AppendAllText`
with an empty buffer.

Adds regression tests (inverting the ones that ensured deletion).